+++ tests/tests_pytorch/deprecated_api/test_remove_1-7.py
@@ -20,7 +20,7 @@ from unittest.mock import Mock
 import pytest
 import torch
 
-from pytorch_lightning import Callback, Trainer
+from pytorch_lightning import Trainer
 from pytorch_lightning.demos.boring_classes import BoringModel
 from pytorch_lightning.overrides.distributed import IndexBatchSamplerWrapper
 from pytorch_lightning.plugins.environments import (
@@ -31,31 +31,7 @@ from pytorch_lightning.plugins.environments import (
     TorchElasticEnvironment,
 )
 from pytorch_lightning.strategies import SingleDeviceStrategy
-from tests_pytorch.deprecated_api import _soft_unimport_module
 from tests_pytorch.plugins.environments.test_lsf_environment import _make_rankfile
-
-
-def test_v1_7_0_on_interrupt(tmpdir):
-    class HandleInterruptCallback(Callback):
-        def on_keyboard_interrupt(self, trainer, pl_module):
-            print("keyboard interrupt")
-
-    model = BoringModel()
-    handle_interrupt_callback = HandleInterruptCallback()
-
-    trainer = Trainer(
-        callbacks=[handle_interrupt_callback],
-        max_epochs=1,
-        limit_val_batches=0.1,
-        limit_train_batches=0.2,
-        enable_progress_bar=False,
-        logger=False,
-        default_root_dir=tmpdir,
-    )
-    with pytest.deprecated_call(
-        match="The `on_keyboard_interrupt` callback hook was deprecated in v1.5 and will be removed in v1.7"
-    ):
-        trainer.fit(model)
 
 
 class BoringCallbackDDPSpawnModel(BoringModel):
@@ -97,15 +73,6 @@ def test_v1_7_0_deprecate_on_post_move_to_device(tmpdir):
         match=r"Method `on_post_move_to_device` has been deprecated in v1.5 and will be removed in v1.7"
     ):
         trainer.fit(model)
-
-
-def test_v1_7_0_deprecate_parameter_validation():
-
-    _soft_unimport_module("pytorch_lightning.core.decorators")
-    with pytest.deprecated_call(
-        match="Using `pytorch_lightning.core.decorators.parameter_validation` is deprecated in v1.5"
-    ):
-        from pytorch_lightning.core.decorators import parameter_validation  # noqa: F401
 
 
 def test_v1_7_0_deprecated_slurm_job_id():